requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langgraph==0.2.20
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
pydantic==2.11.7
datetime
pytz==2023.3
requests==2.31.0
aiofiles==23.2.0